Drop "obj" target
It was introduced to make object files depend on $(BUILDDIR) so that creating it
could be moved into separate rule, executed once.  That was a bad idea, because
"obj" had to be phony, thus making all targets that depend on it obsolete
between runs of make.  This change introduces order-only rule that makes object
files depend on $(BUILDDIR) without obsoleting other targets.