Implement key binding logic
This code is *heavily* based on work I previously did for aerc. Since
I'm the author of that code, and therefore the copyright holder, I'm
able to heavily re-use logic. Of course, because of the differences
between termbox and SDL's event logic and key naming features, there's
some significant differences in how binds are handled, and how keys are
named.