Fix truncated path list from stdin
It turns out, the paths from stdin weren't being truncated. The issue
was with the use of the SDL event queue for passing the paths back to
imv's main thread.

The events were being pushed correctly, but due to a workaround for a
bug in SDL, instead of flushing the contents of the event queue, we now
ignore window events until all events already in the queue have been
handled.

This fixes #141